requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
google==3.0.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud==0.34.0
google-cloud-vision==3.5.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytesseract==0.3.10
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
